Type checking.
As of now all operators except for notable exceptions should be type
checked. These exceptions are operators like CALL as calls themselves
are, in the case of the left operands, checked for a matching function.

Moreover, I have now fixed an issue with comments that has been plaging
the compiler since day 1. The thing that comes after the end of a
comment will be handled very weirdly but that is no longer the case and
comments should not affect the lexing process.

To add on to this I made a lot of smaller changes like adding AST_IMPL
printing support as well as a definition that goes along with it to
limit not print it's body due to the size of the body being equal to
a small module.