Updated module importing, closing #28
Modules are now parsed to completition, one at a time

Removed 'as' (BIT_CAST) operator

Added structs for every AST type

Added symbols as an AST type and modified the defintion of a symbol

All IDs are parsed to symbols now instead which should then later be
resolved. This means that there always only ever one version of a
declaration in the AST while multiple symbols can point to it.